remove old retries and service
