+++ src/sinks/aws_cloudwatch_logs/mod.rs
@@ -6,11 +6,10 @@ use crate::{
     region::RegionOrEndpoint,
     sinks::util::{
         encoding::{EncodingConfig, EncodingConfiguration},
-        retries2::{FixedRetryPolicy, RetryLogic},
-        rusoto,
-        service2::{TowerCompat, TowerRequestConfig, TowerRequestSettings},
-        BatchConfig, BatchSettings, Compression, EncodedLength, PartitionBatchSink,
-        PartitionBuffer, PartitionInnerBuffer, VecBuffer,
+        retries::{FixedRetryPolicy, RetryLogic},
+        rusoto, BatchConfig, BatchSettings, Compression, EncodedLength, PartitionBatchSink,
+        PartitionBuffer, PartitionInnerBuffer, TowerCompat, TowerRequestConfig,
+        TowerRequestSettings, VecBuffer,
     },
     template::Template,
     topology::config::{DataType, SinkConfig, SinkContext},
