+++ src/sinks/clickhouse.rs
@@ -3,9 +3,8 @@ use crate::{
     sinks::util::{
         encoding::{EncodingConfigWithDefault, EncodingConfiguration},
         http::{Auth, BatchedHttpSink, HttpClient, HttpRetryLogic, HttpSink},
-        retries2::{RetryAction, RetryLogic},
-        service2::TowerRequestConfig,
-        BatchConfig, BatchSettings, Buffer, Compression,
+        retries::{RetryAction, RetryLogic},
+        BatchConfig, BatchSettings, Buffer, Compression, TowerRequestConfig,
     },
     tls::{TlsOptions, TlsSettings},
     topology::config::{DataType, SinkConfig, SinkContext, SinkDescription},
