+++ src/sinks/util/retries.rs
@@ -1,10 +1,14 @@
-use super::service::Elapsed;
 use crate::Error;
-use futures::{FutureExt, TryFutureExt};
-use futures01::{try_ready, Async, Future, Poll};
-use std::{cmp, time::Duration};
-use tokio::time::delay_for;
-use tower::retry::Policy;
+use futures::FutureExt;
+use std::{
+    cmp,
+    future::Future,
+    pin::Pin,
+    task::{Context, Poll},
+    time::Duration,
+};
+use tokio::time::{delay_for, Delay};
+use tower03::{retry::Policy, timeout::error::Elapsed};
 
 pub enum RetryAction {
     /// Indicate that this request should be retried with a reason
@@ -37,7 +41,7 @@ pub struct FixedRetryPolicy<L> {
 }
 
 pub struct RetryPolicyFuture<L: RetryLogic> {
-    delay: Box<dyn Future<Item = (), Error = ()> + Send>,
+    delay: Delay,
     policy: FixedRetryPolicy<L>,
 }
 
@@ -75,8 +79,7 @@ impl<L: RetryLogic> FixedRetryPolicy<L> {
 
     fn build_retry(&self) -> RetryPolicyFuture<L> {
         let policy = self.advance();
-        let delay = delay_for(policy.backoff());
-        let delay = Box::new(async move { Ok(delay.await) }.boxed().compat());
+        let delay = delay_for(self.backoff());
 
         debug!(message = "retrying request.", delay_ms = %self.backoff().as_millis());
         RetryPolicyFuture { delay, policy }
@@ -142,13 +145,16 @@ where
     }
 }
 
-impl<L: RetryLogic> Future for RetryPolicyFuture<L> {
-    type Item = FixedRetryPolicy<L>;
-    type Error = ();
+// Safety: `L` is never pinned and we use no unsafe pin projections
+// therefore this safe.
+impl<L: RetryLogic> Unpin for RetryPolicyFuture<L> {}
 
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        try_ready!(self.delay.poll().map_err(|_| unreachable!()));
-        Ok(Async::Ready(self.policy.clone()))
+impl<L: RetryLogic> Future for RetryPolicyFuture<L> {
+    type Output = FixedRetryPolicy<L>;
+
+    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        futures::ready!(self.delay.poll_unpin(cx));
+        Poll::Ready(self.policy.clone())
     }
 }
 
@@ -178,52 +184,19 @@ impl RetryAction {
     }
 }
 
-// Disabling these tests because somehow I triggered a rustc
-// bug where we can only have one assert_eq in play.
-//
-// rustc issue: https://github.com/rust-lang/rust/issues/71259
-#[cfg(feature = "disabled")]
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::test_util::trace_init;
-    use futures01::Future;
     use std::{fmt, time::Duration};
-    use tokio01_test::{assert_err, assert_not_ready, assert_ready, clock};
-    use tower::{retry::Retry, Service};
-    use tower_test01::{assert_request_eq, mock};
+    use tokio::time;
+    use tokio_test::{assert_pending, assert_ready_err, assert_ready_ok, task};
+    use tower03::retry::RetryLayer;
+    use tower_test03::{assert_request_eq, mock};
 
-    #[test]
-    fn service_error_retry() {
-        clock::mock(|clock| {
-            trace_init();
-
-            let policy = FixedRetryPolicy::new(
-                5,
-                Duration::from_secs(1),
-                Duration::from_secs(10),
-                SvcRetryLogic,
-            );
-
-            let (service, mut handle) = mock::pair();
-            let mut svc = Retry::new(policy, service);
-
-            assert_ready!(svc.poll_ready());
-
-            let mut fut = svc.call("hello");
-            assert_request_eq!(handle, "hello").send_error(Error(true));
-            assert_not_ready!(fut.poll());
-
-            clock.advance(Duration::from_secs(2));
-            assert_not_ready!(fut.poll());
-
-            assert_request_eq!(handle, "hello").send_response("world");
-            assert_eq!(fut.wait().unwrap(), "world");
-        });
-    }
-
-    #[test]
-    fn service_error_no_retry() {
+    #[tokio::test]
+    async fn service_error_retry() {
+        time::pause();
         trace_init();
 
         let policy = FixedRetryPolicy::new(
@@ -233,44 +206,69 @@ mod tests {
             SvcRetryLogic,
         );
 
-        let (service, mut handle) = mock::pair();
-        let mut svc = Retry::new(policy, service);
+        let (mut svc, mut handle) = mock::spawn_layer(RetryLayer::new(policy));
 
-        assert_ready!(svc.poll_ready());
+        assert_ready_ok!(svc.poll_ready());
 
-        let mut fut = svc.call("hello");
-        assert_request_eq!(handle, "hello").send_error(Error(false));
-        assert_err!(fut.poll());
+        let fut = svc.call("hello");
+        let mut fut = task::spawn(fut);
+
+        assert_request_eq!(handle, "hello").send_error(Error(true));
+
+        assert_pending!(fut.poll());
+
+        time::advance(Duration::from_secs(2)).await;
+        assert_pending!(fut.poll());
+
+        assert_request_eq!(handle, "hello").send_response("world");
+        assert_eq!(fut.await.unwrap(), "world");
     }
 
-    #[test]
-    fn timeout_error() {
-        clock::mock(|clock| {
-            trace_init();
+    #[tokio::test]
+    async fn service_error_no_retry() {
+        trace_init();
 
-            let policy = FixedRetryPolicy::new(
-                5,
-                Duration::from_secs(1),
-                Duration::from_secs(10),
-                SvcRetryLogic,
-            );
+        let policy = FixedRetryPolicy::new(
+            5,
+            Duration::from_secs(1),
+            Duration::from_secs(10),
+            SvcRetryLogic,
+        );
 
-            let (service, mut handle) = mock::pair();
-            let mut svc = Retry::new(policy, service);
+        let (mut svc, mut handle) = mock::spawn_layer(RetryLayer::new(policy));
 
-            assert_ready!(svc.poll_ready());
+        assert_ready_ok!(svc.poll_ready());
 
-            let mut fut = svc.call("hello");
-            assert_request_eq!(handle, "hello")
-                .send_error(super::super::service::Elapsed::default());
-            assert_not_ready!(fut.poll());
+        let mut fut = task::spawn(svc.call("hello"));
+        assert_request_eq!(handle, "hello").send_error(Error(false));
+        assert_ready_err!(fut.poll());
+    }
 
-            clock.advance(Duration::from_secs(2));
-            assert_not_ready!(fut.poll());
+    #[tokio::test]
+    async fn timeout_error() {
+        time::pause();
+        trace_init();
 
-            assert_request_eq!(handle, "hello").send_response("world");
-            assert_eq!(fut.wait().unwrap(), "world");
-        });
+        let policy = FixedRetryPolicy::new(
+            5,
+            Duration::from_secs(1),
+            Duration::from_secs(10),
+            SvcRetryLogic,
+        );
+
+        let (mut svc, mut handle) = mock::spawn_layer(RetryLayer::new(policy));
+
+        assert_ready_ok!(svc.poll_ready());
+
+        let mut fut = task::spawn(svc.call("hello"));
+        assert_request_eq!(handle, "hello").send_error(tower03::timeout::error::Elapsed::new());
+        assert_pending!(fut.poll());
+
+        time::advance(Duration::from_secs(2)).await;
+        assert_pending!(fut.poll());
+
+        assert_request_eq!(handle, "hello").send_response("world");
+        assert_eq!(fut.await.unwrap(), "world");
     }
 
     #[test]
